requirements.txt
paho-mqtt==2.1.0
PyQt5==5.15.10
pyqtgraph==0.13.7